refactor: 패키지 분리 utils 컴포넌트 추가 및 리팩토링
