+++ backend/src/main/java/com/doodledoodle/backend/dictionary/repository/JpaDictionaryRepository.java
@@ -3,5 +3,4 @@ package com.doodledoodle.backend.dictionary.repository;
 import com.doodledoodle.backend.dictionary.entity.Dictionary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaDictionaryRepository extends JpaRepository<Dictionary, Long>, DictionaryRepository {
-}
+public interface JpaDictionaryRepository extends JpaRepository<Dictionary, Long>, DictionaryRepository {}
