+++ backend/src/main/java/com/doodledoodle/backend/dictionary/service/DictionaryService.java
@@ -7,7 +7,7 @@ import com.doodledoodle.backend.dictionary.repository.DictionaryRepository;
 import com.doodledoodle.backend.global.EntityLoader;
 import com.doodledoodle.backend.global.exception.EntityNotFoundException;
 import com.doodledoodle.backend.result.dto.collection.DictionaryMap;
-import com.doodledoodle.backend.util.RandomGenerator;
+import com.doodledoodle.backend.utils.RandomGenerator;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
