+++ backend/src/main/java/com/doodledoodle/backend/global/kafka/KafkaConsumer.java
@@ -2,4 +2,6 @@ package com.doodledoodle.backend.global.kafka;
 
 public interface KafkaConsumer<T> {
     void consume(String message);
+
+    T readMessage(String message);
 }
