+++ backend/src/main/java/com/doodledoodle/backend/utils/RandomGenerator.java
@@ -1,10 +1,13 @@
-package com.doodledoodle.backend.util;
+package com.doodledoodle.backend.utils;
 
+import lombok.AccessLevel;
+import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Component;
 
 import java.util.Random;
 
 @Component
+@FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 public class RandomGenerator {
     private final Random random;
